+++ {{cookiecutter.repo_name}}/settings/__init__.py
@@ -9,6 +9,4 @@ if "test" in sys.argv:
     print("Try: \033[1;33mpy.test\033[0m")
     sys.exit(0)
 
-from .development import Development  # noqa
-from .production import Production  # noqa
-from .testing import Testing  # noqa
+# Removed Future Import